+++ R/check_odd_k.R
@@ -1,3 +1,0 @@
-check_odd_k = function(k) {
-  if(k %% 2 != 1) stop("k must be an odd positive integer.")
-}