+++ R/check_one_attribute.R
@@ -1,4 +1,0 @@
-check_one_attribute = function(x) {
-  if(length(x) > 1) warning("Only first attribute used.")
-  return(x[1])
-}