+++ R/focal2.R
@@ -1,65 +1,112 @@
 #' Apply a focal filter on a raster
 #'
-#' Applies a focal filter with neighborhood size \code{k}*\code{k} on a raster (class \code{stars}).
+#' Applies a focal filter with neighborhood size \code{k}*\code{k} on a raster (class \code{stars}), using C code.
 #'
 #' @param	x	A raster (class \code{stars}) with two dimensions: \code{x} and \code{y}, i.e., a single-band raster
-#' @param k Neighborhood size around the focal cell. For example, \code{k=3} implies a neighborhood of size 3*3. Must be an odd positive integer.
-#' @param fun A function to be applied on each 3x3 neighborhood. The function needs to accepts a vector of length 9 and return a vector of length 1
+#' @param w Weights matrix defining the neighborhood size around the focal cell, as well as the weights. For example, \code{matrix(1,3,3)} implies a neighborhood of size 3*3 with equal weights of 1 for all cells. The matrix must be square, with odd number of rows and columns.
+#' @param fun A function to aggregate the resulting values for each neighborhood. Possible values are: \code{"mean"}, \code{"sum"}, \code{"min"}, \code{"max"}. The default is \code{"mean"}, i.e., the resulting values per neighborhood are \emph{averaged} before being assigned to the new focal cell value.
+#' @param weight_fun A function which is applied for each pair of cell value and weight value. Possible values are: \code{"+"}, \code{"-"}, \code{"*"}, \code{"/"}. The default is \code{"*"}, i.e., each cell value is \emph{multiplied} by the respective weight.
+#' @param na.rm Should \code{NA} values be removed before applying function? Dafault is \code{TRUE}
 #' @param mask If \code{TRUE}, pixels with \code{NA} in the input are set to \code{NA} in the output as well, i.e., the output is "masked" with the input (default \code{FALSE})
-#' @param ... Further arguments passed to \code{fun}
+#' @param na_flag Value use to mark \code{NA} values in C code. This should be set to a value which is guaranteed to be absent from the raster (default is \code{-9999}).
 #' @return The filtered \code{stars} raster
 #'
 #' @note The raster is "padded" with one more row/column of \code{NA} values on all sides, so that the neigborhood of the outermost rows and columns is still a complete 3x3 neighborhood. Those rows and columns are removed from the filtered result before returning it.
+#' @references The function interface was inspired by function \code{raster::focal}. The C code for this function is a modified and expanded version of the C function named \code{applyKernel} included with package \code{spatialfil}.
 #'
 #' @examples
 #' # Small example
 #' data(dem)
-#' dem1 = focal2(dem, 3, mean, na.rm = TRUE)
-#' dem2 = focal2(dem, 3, min, na.rm = TRUE)
-#' dem3 = focal2(dem, 3, max, na.rm = TRUE)
-#' r = c(dem, round(dem1, 1), dem2, dem3, along = 3)
-#' r = st_set_dimensions(r, 3, values = c("input", "mean", "min", "max"))
+#' dem1 = focal2(dem, matrix(1, 3, 3), "mean")
+#' r = c(dem, round(dem1, 1), along = 3)
+#' r = st_set_dimensions(r, 3, values = c("input", "mean"))
 #' plot(r, text_values = TRUE, breaks = "equal", col = terrain.colors(10))
 #'
 #' \dontrun{
+#'
 #' # Larger example
 #' data(carmel)
-#' carmel1 = focal2(carmel, 3, mean, na.rm = TRUE)
-#' carmel2 = focal2(carmel, 9, mean, na.rm = TRUE)
-#' carmel3 = focal2(carmel, 15, mean, na.rm = TRUE)
-#' r = c(carmel, carmel1, carmel2, carmel3, along = 3)
-#' r = st_set_dimensions(r, 3, values = c("input", "k=3", "k=5", "k=7"))
-#' plot(r, breaks = "equal", col = terrain.colors(100))
+#' carmel1 = focal2(carmel, matrix(1, 15, 15), "mean")
+#' r = c(carmel, round(carmel1, 1), along = 3)
+#' r = st_set_dimensions(r, 3, values = c("input", "mean"))
+#' plot(r, breaks = "equal", col = terrain.colors(10))
+#'
 #' }
 #'
 #' @export
 
 # Apply focal filter
-focal2 = function(x, k = 3, fun, mask = FALSE, ...) {
-
-  # Number of steps in each direction beyond focal cell
-  steps = (k - 1) / 2
+focal2 = function(x, w, fun = "mean", weight_fun = "*", na.rm = FALSE, mask = FALSE, na_flag = -9999) {
 
   # Make template
   template = x
 
+  # Checks
+  stopifnot(is(w, "matrix"))
+  if(any(is.na(w))) { stop("weight matrix 'w' includes 'NA's") }
+  if(!nrow(w) == ncol(w)) { stop("weight matrix is not rectangular") }
+  stopifnot(fun %in% c("mean", "sum", "min", "max"))
+  stopifnot(weight_fun %in% c("+", "-", "*", "/"))
+
+  # Number of extra lines
+  steps = (nrow(w) - 1) / 2
+
   # To matrix
-  input = layer_values_matrix(template)  # Checks take place here
+  input = layer_to_matrix(template)  # Checks take place here
   input = matrix_extend(input, n = steps)
 
-  # Apply filter
-  output = matrix(NA, nrow = nrow(input), ncol = ncol(input))
-  for(i in (1+steps):(nrow(input)-steps)) {
-    for(j in (1+steps):(ncol(input)-steps)) {
-      v = matrix_get_neighbors(m = input, pos = c(i, j), k = k)
-      output[i, j] = fun(v, ...)
-    }
-  }
+  # Matrix dimensions
+  nrows = nrow(input)
+  ncols = ncol(input)
 
-  # Back to raster
+  # Index matrix for the kernel
+  kindex = NULL
+  for(n in -steps:steps)  # Index for columns
+    for(m in -steps:steps)  # Index for rows
+      kindex = c(kindex, n*ncols + m)
+
+  # To vector
+  input_vector = as.vector(t(input))
+
+  # Replace 'NA' with 'na_flag'
+  input_vector2 = input_vector
+  input_vector2[is.na(input_vector2)] = na_flag
+  output_vector2 = rep(na_flag, length(input_vector2))
+
+  # Encode 'fun' and 'weight_fun' to integer
+  if(fun == "mean") fun = 1
+  if(fun == "sum") fun = 2
+  if(fun == "min") fun = 3
+  if(fun == "max") fun = 4
+  if(weight_fun == "+") weight_fun = 1
+  if(weight_fun == "-") weight_fun = 2
+  if(weight_fun == "*") weight_fun = 3
+  if(weight_fun == "/") weight_fun = 4
+
+  # Apply filter
+  result = .C(
+    "focal2c",
+    as.double(input_vector2),
+    as.integer(nrows),
+    as.integer(ncols),
+    as.double(t(w)),
+    as.integer(steps),
+    as.integer(kindex),
+    as.integer(na.rm),
+    as.double(na_flag),
+    as.integer(fun),
+    as.integer(weight_fun),
+    as.double(output_vector2)
+  )
+  output_vector = result[[length(result)]]
+
+  # Back to 'stars'
+  if(mask) output_vector[is.na(input_vector)] = NA
+  output_vector[output_vector == na_flag] = NA
+  output = matrix(output_vector, nrow(input), ncol(input), byrow = TRUE)
   output = matrix_trim(output, n = steps)
-  if(mask) output[is.na(input)] = NA
-  template[[1]] = t(output)
+  output = t(output)
+  template[[1]] = output
 
   # Return
   return(template)