+++ R/starsExtra-package.R
@@ -1,4 +1,10 @@
 #' @keywords internal
+#' @import sf
 #' @import stars
+#' @importFrom methods is
 "_PACKAGE"
 
+## usethis namespace: start
+#' @useDynLib starsExtra, .registration = TRUE
+## usethis namespace: end
+NULL