+++ src/starsExtra_init.c
@@ -1,0 +1,16 @@
+#include <stdlib.h> // for NULL
+#include <R_ext/Rdynload.h>
+
+/* .C calls */
+extern void focal2c(void *, void *, void *, void *, void *, void *, void *, void *, void *, void *, void *);
+
+static const R_CMethodDef CEntries[] = {
+    {"focal2c", (DL_FUNC) &focal2c, 11},
+    {NULL, NULL, 0}
+};
+
+void R_init_starsExtra(DllInfo *dll)
+{
+    R_registerRoutines(dll, CEntries, NULL, NULL, NULL);
+    R_useDynamicSymbols(dll, FALSE);
+}