+++ NEWS.md
@@ -1,4 +1,12 @@
 # starsExtra 0.0.1
 
 * Initial version
+* Added 'focal2r' and 'focal2' functions
 * Added 'focal2' function
+* Added 'flowlength' function
+* Added 'detrend' function
+
+# To do
+* Add "min", "max" and "sum" options
+* Use 'NA' flag in C code
+* Add 'circularweights' function